refactor(#121): Phase 2 - Split useApp.ts and extract groupUpdatesIntoMessages
Split monolithic useApp.ts (953 lines) into focused hooks:
- useProjects.ts (131 lines) - Project CRUD, reordering
- useSessions.ts (325 lines) - Session CRUD, selection, state
- useAgentStatus.ts (248 lines) - Running/processing state, agent actions
- useAppSubscriptions.ts (254 lines) - IPC event handlers
- useApp.ts (316 lines) - Composition layer combining hooks

Extract groupUpdatesIntoMessages from ChatView.tsx:
- ChatView.tsx reduced from ~1028 to 644 lines
- messageGrouping.ts (456 lines) - Pure utility for message grouping
- sessionUpdates.ts (78 lines) - Session update merge utility

Maintains same external interface from useApp.ts - no breaking changes.
All 367 tests pass.
